% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{boot_dissim}
\alias{boot_dissim}
\title{Bootstrap dissimilarity}
\usage{
boot_dissim(lst_vect, index, by, replace, size_short, n_boot, n_cpu)
}
\arguments{
\item{lst_vect}{A list with two character vectors of interactions.}

\item{index}{The name of the desired index, e.g.: "whittaker", "jaccard",
"sorensen". Character type.}

\item{by}{Integer. Number of interactions used to increase gradually the
sampled vectors of interactions until all observations are sampled. If \code{by}
is too small (e.g. 1) then the computation time is very long depending on
your total number of interactions from which samples are taken. As a rule
of thum set \code{by} to maybe 5-10\\\% of your total interactions.}

\item{replace}{Should sampling be with replacement? Passed to
\code{\link[base]{sample}}.}

\item{size_short}{Should the total sample size be set by the length of the
shortest vector from \code{lst_vect}? Logic type (\code{TRUE} or \code{FALSE}).}

\item{n_boot}{Number of desired bootstraps (100 can be enough).}

\item{n_cpu}{Number of CPU-s to use for parallel processing.}
}
\value{
A matrix of dissimilarities. The number of columns corresponds to
\code{n_boot} (number of bootstraps). The names of the rows store the sample
size of each sampled chunk. The final value corresponds to the length of
the shortest vector (when \code{size_short = TRUE}), or to the length of the
longest vector (when \code{size_short = FALSE}).
}
\description{
Bootstrap dissimilarity between two vectors of interactions (or
species names).
}
\examples{

library(bootdissim)
library(bipartite)

net1 <- reshape_net(vazarr, seed = 1)
net2 <- reshape_net(vazcer, seed = 1)

vect <- get_interactions(net1, net2)

tst <- boot_dissim(lst_vect = vect,
                   index = "whittaker",
                   by = 50,
                   replace = TRUE,
                   size_short = FALSE,
                   n_boot = 10,
                   n_cpu = 2)

}
