% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{boot_dissim_once}
\alias{boot_dissim_once}
\title{Compute dissimilarity for a sample}
\usage{
boot_dissim_once(vect_short, vect_long, index, size, by, replace, seed, n)
}
\arguments{
\item{vect_short}{The shorter vectors of interactions.}

\item{vect_long}{The longer vectors of interactions.}

\item{index, size, by, replace}{See \code{\link[bootdissim]{boot_dissim}}.}

\item{seed}{Passed to \code{\link[base]{set.seed}}. Set seed to get
reproducible random results. Is give by the iterator of the loop during the
parallel processing in \code{\link[bootdissim]{boot_dissim}}.}

\item{n}{Passed from \code{\link[bootdissim]{boot_dissim}}. Is constant
across all bootstrap iterations.}
}
\value{
A vector of dissimilarities. The length of this vector corresponds to
\code{n}.
}
\description{
Compute dissimilarity between two vectors of interactions (or
species names) for a single run (sample). You will rarely use this function
alone. It was designed to be executed in parallel by
\code{\link[bootdissim]{boot_dissim}}. See more details there.
}
