% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootdissim.r
\docType{package}
\name{bootdissim}
\alias{bootdissim}
\title{Bootstrap dissimilarities between networks}
\description{
R package for resampling network dissimilarities Given two
networks of plant - pollinator interactions one can compute and visualize
how the dissimilarity between these two networks changes with number of
sampled interactions.
}
\author{
Valentin Stefan, Tiffany Knight
}
