% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.r
\name{get_abc}
\alias{get_abc}
\title{Compute turnover components.}
\usage{
get_abc(vect)
}
\arguments{
\item{vect}{A list of two vectors with species names or interactions.}
}
\value{
A list with the 3 components a, b, c. The elements of the list are
numeric vectors of length 1.
}
\description{
Compute the a, b, c turnover components. Given two vectors, v1
and v2, of species names or interactions, component "a" is the number of
elements (species or interactions) present in both vectors v1 and v2, "b"
is the number of elements present only in v1 and "c" only in v2. See the
section "Beta diversity estimates" in  Novotny, V. (2009). Beta diversity
of plant–insect food webs in tropical forests: a conceptual framework.
Insect Conservation and Diversity, 2(1), 5-9.
}
\examples{

library(bootdissim)
library(bipartite)

net1 <- reshape_net(vazarr, seed = 1)
net2 <- reshape_net(vazcer, seed = 1)
vect <- get_interactions(net1, net2)
abc <- get_abc(vect)

}
