% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.r
\name{get_beta}
\alias{get_beta}
\title{Compute dissimilarity index.}
\usage{
get_beta(abc, index)
}
\arguments{
\item{abc}{A list with the turnover components a, b and c.}

\item{index}{The name of the desired index, e.g.: "whittaker", "jaccard",
"sorensen". Character type.}
}
\value{
A dissimilarity index. Numeric vector of length one.
}
\description{
Compute various dissimilarity indices.
}
\examples{

library(bootdissim)
library(bipartite)

net1 <- reshape_net(vazarr, seed = 1)
net2 <- reshape_net(vazcer, seed = 1)
vect <- get_interactions(net1, net2)
abc <- get_abc(vect)
get_beta(abc, index = "whittaker")
get_beta(abc, index = "jaccard")
get_beta(abc, index = "sorensen")

}
