% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_interactions}
\alias{get_interactions}
\title{Get interactions}
\usage{
get_interactions(net1, net2)
}
\arguments{
\item{net1, net2}{Two data frames of interactions, for example constructed
with \code{\link[bootdissim]{reshape_net}}. Each data frame must have only
two columns/variables, e.g. first column for plant names, second column for
insect names.}
}
\value{
Returns a list of two character vectors.
}
\description{
Helper function which constructs two vectors of species
interactions (e.g. plant - pollinator) for two network interactions
provided as data frames.
}
\examples{

library(bootdissim)
library(bipartite)

net1 <- reshape_net(vazarr, seed = 1)
net2 <- reshape_net(vazcer, seed = 1)

vect <- get_interactions(net1, net2)

}
