% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.r
\name{get_stats_gg}
\alias{get_stats_gg}
\title{Prepare bootstrapped results for \code{ggplot} friendly data format.}
\usage{
get_stats_gg(x, probs = c(0.025, 0.975))
}
\arguments{
\item{x}{A matrix of bootstrapped dissimilarities from
\code{\link[bootdissim]{boot_dissim}}}

\item{probs}{A numeric vector of two probabilities in the interval \verb{[0, 1]}.
Used for building the lower and upper bounds of the confidence intervals.
Defaults to \code{c(0.025, 0.975)}, which corresponds to a 95\\\% confidence
interval.}
}
\value{
A list of two data frames to be used with
\code{\link[ggplot2]{ggplot}}. First one, \code{stats_df} is a 4 columns data
frame, containing the sample size/number of sampled interactions
(\code{spl_size}) at each step and the corresponding mean dissimilarity (\code{mean})
together with its bootstrap confidence interval limits (\code{ci_low}, \code{ci_up}).
The second one, \code{lines_df}, contains all the bootstrapped dissimilarities
(\code{value}) at each bootstrap iteration (\code{simulation_id}) and its
corresponding sample size/sampled interactions (\code{spl_size}). Can be used
for enhancing visual effect when plotting the mean bootstrap values.
}
\description{
Helper function to convert the bootstrapped dissimilarities from
matrix format to \code{ggplot2} friendly data format.
}
