% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{reshape_net}
\alias{reshape_net}
\title{Transform network (web) matrix to data table}
\usage{
reshape_net(net, abund = TRUE, seed)
}
\arguments{
\item{net}{A network matrix (a web), e.g. \code{Safariland} from
\code{\link{bipartite}}.}

\item{abund}{Logic type (\code{TRUE} or \code{FALSE}). Default is \code{TRUE}, that is
abundance information is considered and an interaction is repeated as many
times it was observed. If \code{FALSE} then it returns only the unique
interactions.}

\item{seed}{A single integer value. Passed to \code{\link[base]{set.seed}}.
Set seed to get reproducible random results.}
}
\value{
Returns a data frame.
}
\description{
Helper function to transform a network matrix (web) to a data
frame, where each row represents an interaction between a low level and
high level species (e.g. plant - pollinator).
}
\examples{

library(bootdissim)
library(bipartite)

# Considering abundance information
Safariland_dt <- reshape_net(Safariland, abund = TRUE, seed = 1)

# Without abundance information
Safariland_dt <- reshape_net(Safariland, abund = FALSE, seed = 1)

}
