% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{sample_vectors}
\alias{sample_vectors}
\title{Sample from two vectors}
\usage{
sample_vectors(v1, v2, size, by, seed, replace)
}
\arguments{
\item{v1, v2}{Vectors of interactions from which to sample.}

\item{size}{Integer giving the total sample size which, for example, can be
the length of the shortest vector of interactions or of the longest one.
Passed to \code{\link[base]{sample}}.}

\item{by}{Integer. Number of interactions used to increase gradually the
sampled vectors of interactions until all observations are sampled (given
in the \code{size} argument).}

\item{seed}{A single integer value. Passed to \code{\link[base]{set.seed}}.
Set seed to get reproducible random results.}

\item{replace}{Should sampling be with replacement? Passed to
\code{\link[base]{sample}}.}
}
\value{
A list of two lists of character vectors containing interactions.
}
\description{
Prepares a list of sample values from two given vectors \code{v1} &
\code{v2} for the bootstrapping workflow. This function is important for
internal use.
}
